refactor(explore): modularize explore page components and optimize loading UX
- Restructure `src/app/(app)/explore/page.tsx` by moving inline declarations into dedicated standalone components under `src/components/explore/`:
  - `SidebarLoadingSkeleton`
  - `ExploreMetricsDashboard`
  - `SearchParamSync`
  - `VisionAnalysisCard`
  - `ExploreDesktopListInterventions`
- Fix selection delay by firing `onFeatureSelected` synchronously at the start of feature selection in `feature_selection.ts`, ensuring immediate transition feedback.
- Enhance skeleton loading styling with standard Tailwind grey colors (`bg-neutral-200` / `bg-neutral-300` and dark mode equivalents) and pulse animations to make the loader clearly visible.
